refactor(react): remove composed modal v10 aand move next to latest
