+++ packages/react/src/components/ComposedModal/ComposedModal-test.js
@@ -6,212 +6,12 @@
  */
 
 import React from 'react';
-import { shallow, mount } from 'enzyme';
-import Button from '../Button';
-import {
-  ComposedModal,
-  ModalHeader,
-  ModalBody,
-  ModalFooter,
-} from '../ComposedModal';
-import InlineLoading from '../InlineLoading';
+import { mount } from 'enzyme';
+import ComposedModal from './ComposedModal';
+import { ModalHeader } from './ModalHeader';
+import { ModalFooter } from './ModalFooter';
 
 const prefix = 'cds';
-
-describe('<ModalHeader />', () => {
-  describe('Renders as expected', () => {
-    const titleWrapper = mount(<ModalHeader title="Something" />);
-    const labelWrapper = mount(<ModalHeader label="Something" />);
-
-    it('does not render title if no title', () => {
-      expect(
-        labelWrapper.find(`.${prefix}--modal-header__heading`).exists()
-      ).toBe(false);
-    });
-
-    it('does not render label if no label', () => {
-      expect(
-        titleWrapper.find(`.${prefix}--modal-header__label`).exists()
-      ).toBe(false);
-    });
-
-    it('renders title if title text', () => {
-      expect(
-        titleWrapper.find(`.${prefix}--modal-header__heading`).exists()
-      ).toBe(true);
-    });
-
-    it('renders label if label text', () => {
-      expect(
-        labelWrapper.find(`.${prefix}--modal-header__label`).exists()
-      ).toBe(true);
-    });
-
-    it('should render with ref', () => {
-      const ref = React.createRef();
-      mount(<ModalHeader title="Something" forwardedRef={ref} />);
-
-      expect(ref.current).toHaveClass(`${prefix}--modal-header`);
-    });
-  });
-});
-
-describe('<ModalBody />', () => {
-  describe('Renders as expected', () => {
-    const wrapper = shallow(
-      <ModalBody className="extra-class">
-        <p>Test</p>
-      </ModalBody>
-    );
-
-    it('renders children as expected', () => {
-      expect(wrapper.find('p').length).toBe(1);
-    });
-
-    it('renders wrapper as expected', () => {
-      expect(wrapper.find(`.${prefix}--modal-content`).length).toBe(1);
-    });
-
-    it('renders extra classes passed in via className', () => {
-      expect(
-        wrapper.find(`.${prefix}--modal-content`).hasClass('extra-class')
-      ).toEqual(true);
-    });
-    it('should render with ref', () => {
-      const ref = React.createRef();
-      mount(
-        <ModalBody className="extra-class" ref={ref}>
-          <p>Test</p>
-        </ModalBody>
-      );
-
-      expect(ref.current).toHaveClass(`${prefix}--modal-content`);
-    });
-  });
-});
-
-describe('<ModalFooter />', () => {
-  describe('Renders as expected', () => {
-    const wrapper = mount(
-      <ModalFooter className="extra-class">
-        <p>Test</p>
-      </ModalFooter>
-    );
-
-    it('renders children as expected', () => {
-      expect(wrapper.find('p').length).toBe(1);
-    });
-
-    it('renders wrapper as expected', () => {
-      expect(wrapper.length).toBe(1);
-    });
-
-    it('renders extra classes passed in via className', () => {
-      expect(wrapper.hasClass('extra-class')).toEqual(true);
-    });
-  });
-
-  describe('Should render buttons only if appropriate prop passed in', () => {
-    const wrapper = shallow(
-      <ModalFooter className="extra-class">
-        <p>Test</p>
-      </ModalFooter>
-    );
-
-    const primaryWrapper = shallow(<ModalFooter primaryButtonText="test" />);
-    const secondaryWrapper = mount(<ModalFooter secondaryButtonText="test" />);
-    const multipleSecondaryWrapper = mount(
-      <ModalFooter
-        secondaryButtons={[
-          {
-            buttonText: <InlineLoading />,
-            onClick: jest.fn(),
-          },
-          {
-            buttonText: 'Cancel',
-            onClick: jest.fn(),
-          },
-        ]}
-      />
-    );
-
-    it('does not render primary button if no primary text', () => {
-      expect(wrapper.find(`.${prefix}--btn--primary`).exists()).toBe(false);
-    });
-
-    it('does not render secondary button if no secondary text', () => {
-      expect(wrapper.find(`.${prefix}--btn--secondary`).exists()).toBe(false);
-    });
-
-    it('renders primary button if primary text', () => {
-      const buttonComponent = primaryWrapper.find(Button);
-      expect(buttonComponent.exists()).toBe(true);
-      expect(buttonComponent.props().kind).toBe('primary');
-    });
-
-    it('renders primary button if secondary text', () => {
-      const buttonComponent = secondaryWrapper.find(Button);
-      expect(buttonComponent.exists()).toBe(true);
-      expect(buttonComponent.props().kind).toBe('secondary');
-    });
-
-    it('correctly renders multiple secondary buttons', () => {
-      const buttonComponents = multipleSecondaryWrapper.find(Button);
-      expect(buttonComponents.length).toEqual(2);
-      expect(buttonComponents.at(0).props().kind).toBe('secondary');
-      expect(buttonComponents.at(1).props().kind).toBe('secondary');
-    });
-  });
-
-  describe('Should render the appropriate buttons when `danger` prop is true', () => {
-    const primaryWrapper = shallow(
-      <ModalFooter primaryButtonText="test" danger />
-    );
-    const secondaryWrapper = mount(
-      <ModalFooter secondaryButtonText="test" danger />
-    );
-    const multipleSecondaryWrapper = mount(
-      <ModalFooter
-        secondaryButtons={[
-          {
-            buttonText: <InlineLoading />,
-            onClick: jest.fn(),
-          },
-          {
-            buttonText: 'Cancel',
-            onClick: jest.fn(),
-          },
-        ]}
-      />
-    );
-
-    it('renders danger button if primary text && danger', () => {
-      const buttonComponent = primaryWrapper.find(Button);
-      expect(buttonComponent.exists()).toBe(true);
-      expect(buttonComponent.props().kind).toBe('danger');
-    });
-
-    it('renders secondary button if secondary text && danger', () => {
-      const buttonComponent = secondaryWrapper.find(Button);
-      expect(buttonComponent.exists()).toBe(true);
-      expect(buttonComponent.prop('kind')).toBe('secondary');
-    });
-
-    it('correctly renders multiple secondary buttons', () => {
-      const buttonComponents = multipleSecondaryWrapper.find(Button);
-      expect(buttonComponents.length).toEqual(2);
-      expect(buttonComponents.at(0).props().kind).toBe('secondary');
-      expect(buttonComponents.at(1).props().kind).toBe('secondary');
-    });
-  });
-
-  it('should render with ref', () => {
-    const ref = React.createRef();
-    mount(<ModalFooter primaryButtonText="test" forwardedRef={ref} />);
-
-    expect(ref.current).toHaveClass(`${prefix}--modal-footer`);
-  });
-});
 
 describe('<ComposedModal />', () => {
   let container;
@@ -230,21 +30,20 @@ describe('<ComposedModal />', () => {
 
   it('renders with a ref', () => {
     const ref = React.createRef();
-    mount(<ComposedModal open forwardedRef={ref} />);
+    mount(<ComposedModal open ref={ref} />);
     expect(ref.current).toHaveClass(`${prefix}--modal`);
-  });
-
-  it('renders with a callbackRef', () => {
-    const ref = jest.fn();
-    const wrapper = mount(<ComposedModal open forwardedRef={ref} />);
-    expect(ref).toHaveBeenCalledWith(wrapper.getDOMNode());
   });
 
   it('changes the open state upon change in props', () => {
     const wrapper = mount(<ComposedModal open />);
-    expect(wrapper.state().open).toEqual(true);
+
+    expect(
+      document.body.classList.contains('cds--body--with-modal-open')
+    ).toEqual(true);
     wrapper.setProps({ open: false });
-    expect(wrapper.state().open).toEqual(false);
+    expect(
+      document.body.classList.contains('cds--body--with-modal-open')
+    ).toEqual(false);
   });
 
   it('should change class of <body> upon open state', () => {
@@ -262,14 +61,6 @@ describe('<ComposedModal />', () => {
     ).toEqual(false);
   });
 
-  it('avoids change the open state upon setting props, unless there the value actually changes', () => {
-    const wrapper = mount(<ComposedModal />);
-    wrapper.setProps({ open: true });
-    wrapper.setState({ open: false });
-    wrapper.setProps({ open: true });
-    expect(wrapper.state().open).toEqual(false);
-  });
-
   it('calls onClick upon user-initiated closing', () => {
     const onClose = jest.fn();
     const wrapper = mount(
@@ -279,7 +70,9 @@ describe('<ComposedModal />', () => {
     );
     const button = wrapper.find(`.${prefix}--modal-close`).first();
     button.simulate('click');
-    expect(wrapper.state().open).toEqual(false);
+    expect(
+      document.body.classList.contains('cds--body--with-modal-open')
+    ).toEqual(false);
     expect(onClose.mock.calls.length).toBe(1);
   });
 
@@ -292,7 +85,9 @@ describe('<ComposedModal />', () => {
     );
     const button = wrapper.find(`.${prefix}--modal-close`).first();
     button.simulate('click');
-    expect(wrapper.state().open).toEqual(true);
+    expect(
+      document.body.classList.contains('cds--body--with-modal-open')
+    ).toEqual(true);
   });
 
   it('should focus on the primary actionable button in ModalFooter by default', () => {
