+++ packages/react/src/components/ComposedModal/ModalHeader-test.js
