+++ packages/react/src/components/ComposedModal/index.js
@@ -5,27 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as FeatureFlags from '@carbon/feature-flags';
-import { ModalHeader as ModalHeaderNext } from './next/ModalHeader';
-import { ModalFooter as ModalFooterNext } from './next/ModalFooter';
-import { default as ComposedModalNext } from './next/ComposedModal';
-import {
-  default as ComposedModalClassic,
-  ModalHeader as ModalHeaderClassic,
-  ModalBody,
-  ModalFooter as ModalFooterClassic,
-} from './ComposedModal';
+import ComposedModal, { ModalBody } from './ComposedModal';
+import { ModalFooter } from './ModalFooter';
+import { ModalHeader } from './ModalHeader';
 
-export const ModalHeader = FeatureFlags.enabled('enable-v11-release')
-  ? ModalHeaderNext
-  : ModalHeaderClassic;
-
-export const ModalFooter = FeatureFlags.enabled('enable-v11-release')
-  ? ModalFooterNext
-  : ModalFooterClassic;
-
-export const ComposedModal = FeatureFlags.enabled('enable-v11-release')
-  ? ComposedModalNext
-  : ComposedModalClassic;
-
-export { ModalBody };
+export { ModalBody, ComposedModal, ModalFooter, ModalHeader };
